Moving the BodyLocation enum to the personnel enums folder
