+++ MekHQ/src/mekhq/campaign/mod/am/HitLocationGen.java
@@ -7,12 +7,12 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 2 of the License, or
  * (at your option) any later version.
- * 
+ *
  * MekHQ is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with MekHQ.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -25,7 +25,7 @@ import java.util.function.Function;
 import java.util.function.IntUnaryOperator;
 
 import mekhq.Utilities;
-import mekhq.campaign.personnel.BodyLocation;
+import mekhq.campaign.personnel.enums.BodyLocation;
 
 /**
  * Home to static methods returning a random hit location given a random integer value generator
@@ -83,11 +83,11 @@ public class HitLocationGen {
         } while((null == entry) || !validCheck.apply(entry.getValue()));
         return entry.getValue();
     }
-    
+
     public static BodyLocation generic(IntUnaryOperator rnd, Function<BodyLocation, Boolean> validCheck) {
         return queryRandomTable(GENERIC_RANDOM_HIT_TABLE, rnd, validCheck);
     }
-    
+
     public static BodyLocation mechAndAsf(IntUnaryOperator rnd, Function<BodyLocation, Boolean> validCheck) {
         return queryRandomTable(MECH_RANDOM_HIT_TABLE, rnd, validCheck);
     }
