+++ MekHQ/src/mekhq/campaign/mod/am/InjuryUtil.java
@@ -31,7 +31,7 @@ import megamek.common.Mech;
 import mekhq.campaign.*;
 import mekhq.campaign.log.MedicalLogger;
 import mekhq.campaign.log.ServiceLogger;
-import mekhq.campaign.personnel.BodyLocation;
+import mekhq.campaign.personnel.enums.BodyLocation;
 import mekhq.campaign.personnel.Injury;
 import mekhq.campaign.personnel.InjuryType;
 import mekhq.campaign.personnel.Person;
