+++ MekHQ/src/mekhq/campaign/personnel/Injury.java
@@ -38,6 +38,7 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
+import mekhq.campaign.personnel.enums.BodyLocation;
 import org.joda.time.DateTime;
 import org.w3c.dom.Node;
