+++ MekHQ/src/mekhq/campaign/personnel/InjuryType.java
@@ -23,6 +23,7 @@ import java.util.*;
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
+import mekhq.campaign.personnel.enums.BodyLocation;
 import org.joda.time.DateTime;
 
 import mekhq.Utilities;
