+++ MekHQ/src/mekhq/campaign/personnel/enums/BodyLocation.java
@@ -7,16 +7,16 @@
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 2 of the License, or
  * (at your option) any later version.
- * 
+ *
  * MekHQ is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with MekHQ.  If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.campaign.personnel;
+package mekhq.campaign.personnel.enums;
 
 import java.util.Arrays;
 import java.util.EnumSet;
@@ -59,12 +59,12 @@ public enum BodyLocation {
             }
         }
     }
-    
+
     /** @return the body location corresponding to the (old) ID */
     public static BodyLocation of(int id) {
         return ((id > 0) && (id < idMap.length)) ? idMap[id] : GENERIC;
     }
-    
+
     /** @return the body location corresponding to the given string */
     public static BodyLocation of(String str) {
         try {
@@ -74,7 +74,7 @@ public enum BodyLocation {
         }
         return valueOf(str.toUpperCase(Locale.ROOT));
     }
-    
+
     public final int id;
     // Includes everything attached to a limb
     public final boolean isLimb;
@@ -87,11 +87,11 @@ public enum BodyLocation {
     private BodyLocation(int id, String readableName) {
         this(id, readableName, false, null);
     }
-    
+
     private BodyLocation(int id, String readableName, boolean isLimb) {
         this(id, readableName, isLimb, null);
     }
-    
+
     private BodyLocation(int id, String readableName, boolean isLimb, BodyLocation parent) {
         this.id = id;
         this.readableName = readableName;
@@ -106,7 +106,7 @@ public enum BodyLocation {
     private void addChildLocation(BodyLocation child) {
         children.add(child);
     }
-    
+
     public boolean isParentOf(BodyLocation child) {
         if(children.contains(child)) {
             return true;
@@ -118,11 +118,11 @@ public enum BodyLocation {
         }
         return false;
     }
-    
+
     public boolean isChildOf(BodyLocation parent) {
         return (null != this.parent) ? (this.parent == parent) || this.parent.isChildOf(parent) : false;
     }
-    
+
     public static final class XMLAdapter extends XmlAdapter<String, BodyLocation> {
         @Override
         public BodyLocation unmarshal(String v) throws Exception {
